Address review: rename rfc822→datetimes, drop redundant Optional wrap
- Rename api/schemas/rfc822.py → api/schemas/datetimes.py now that it no
  longer emits RFC 822 (it's the ISO 8601 serializer + permissive parser).
  Update the three importers.
- Drop the redundant outer `Optional[...]` on `FlexibleDatetimeOpt` fields
  (`FlexibleDatetimeOpt` is already `Annotated[Optional[datetime], ...]`).
